Refactor the write code to make it easier to add new encoders.
