+++ lib/encoders/encoders.h
@@ -1,7 +1,18 @@
 #ifndef ENCODERS_H
 #define ENCODERS_H
 
-class AbstractEncoder {};
+class Fluxmap;
+class SectorSet;
+
+class AbstractEncoder
+{
+public:
+    virtual ~AbstractEncoder() {}
+
+public:
+    virtual std::unique_ptr<Fluxmap> encode(
+        int physicalTrack, int physicalSide, const SectorSet& allSectors) = 0;
+};
 
 #endif
